Requested changes: -Move integration tests to ModelStateValidationTests -formatting -documentation -comments
