+++ docs/usage/options.md
@@ -87,18 +87,7 @@ If you would like to use ASP.NET Core ModelState validation into your controller
 ```c#
 options.ValidateModelState = true;
 ```
-
-# Custom Validators
-
-Attributes can be marked with custom validators which work with ModelState validation.
-
-## IsRequired Attribute
-
-The 'IsRequired' attribute is derived from the 'Required' validator attribute. It accepts a bool to specify if empty strings are allowed on that property. The default for 'AllowEmptyStrings' is false.
-
-The 'IsRequired' attribute can be used to decorate properties allowing the property to be disabled on PATCH requests, making partial patching possible. 
-
-If a PATCH request contains a property assigned the 'IsRequired' attribute, the requirements of the validator are verified against the patched value. When the validator is enabled, that properties value cannot be null and cannot be empty if 'AllowEmptyStrings' is set to false. 
+You will need to use the JsonApiDotNetCore 'IsRequiredAttribute' instead of the built-in 'RequiredAttribute', because it contains modifications to enable partial patching.
 
 ```c#
 public class Person : Identifiable<int>
