+++ src/Examples/JsonApiDotNetCoreExample/Models/Article.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
 using JsonApiDotNetCore.Models;
-using JsonApiDotNetCore.Models.CustomValidators;
 
 namespace JsonApiDotNetCoreExample.Models
 {
