+++ src/Examples/JsonApiDotNetCoreExample/Models/Author.cs
@@ -1,6 +1,5 @@
 using JsonApiDotNetCore.Models;
 using System.Collections.Generic;
-using JsonApiDotNetCore.Models.CustomValidators;
 
 namespace JsonApiDotNetCoreExample.Models
 {
