+++ test/JsonApiDotNetCoreExampleTests/Acceptance/ResourceDefinitions/ResourceDefinitionTests.cs
@@ -613,380 +613,5 @@ namespace JsonApiDotNetCoreExampleTests.Acceptance
             Assert.Equal("You are not allowed to update fields or relationships of locked todo items.", errorDocument.Errors[0].Title);
             Assert.Null(errorDocument.Errors[0].Detail);
         }
-
-        [Fact]
-        public async Task Create_Article_With_IsRequired_Name_Attribute_Succeeds()
-        {
-            // Arrange
-            string name = "Article Title";
-            var context = _fixture.GetService<AppDbContext>();
-            var author = _authorFaker.Generate();
-            context.AuthorDifferentDbContextName.Add(author);
-            await context.SaveChangesAsync();
-
-            var route = "/api/v1/articles";
-            var request = new HttpRequestMessage(new HttpMethod("POST"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    attributes = new Dictionary<string, object>
-                    {
-                        {"name", name}
-                    },
-                    relationships = new Dictionary<string, dynamic>
-                    {
-                        {  "author",  new
-                            {
-                                data = new
-                                {
-                                    type = "authors",
-                                    id = author.StringId
-                                }
-                             }
-                        }
-                    }
-                }
-            };
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            var body = await response.Content.ReadAsStringAsync();
-            Assert.Equal(HttpStatusCode.Created, response.StatusCode);
-
-            var articleResponse = _fixture.GetDeserializer().DeserializeSingle<Article>(body).Data;
-            Assert.NotNull(articleResponse);
-
-            var persistedArticle = await _fixture.Context.Articles
-                .SingleAsync(a => a.Id == articleResponse.Id);
-
-            Assert.Equal(name, persistedArticle.Name);
-        }
-
-        [Fact]
-        public async Task Create_Article_With_IsRequired_Name_Attribute_Empty_Succeeds()
-        {
-            // Arrange
-            string name = string.Empty;
-            var context = _fixture.GetService<AppDbContext>();
-            var author = _authorFaker.Generate();
-            context.AuthorDifferentDbContextName.Add(author);
-            await context.SaveChangesAsync();
-
-            var route = "/api/v1/articles";
-            var request = new HttpRequestMessage(new HttpMethod("POST"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    attributes = new Dictionary<string, object>
-                    {
-                        {"name", name}
-                    },
-                    relationships = new Dictionary<string, dynamic>
-                    {
-                        {  "author",  new
-                            {
-                                data = new
-                                {
-                                    type = "authors",
-                                    id = author.StringId
-                                }
-                             }
-                        }
-                    }
-                }
-            };
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            var body = await response.Content.ReadAsStringAsync();
-            Assert.Equal(HttpStatusCode.Created, response.StatusCode);
-
-            var articleResponse = _fixture.GetDeserializer().DeserializeSingle<Article>(body).Data;
-            Assert.NotNull(articleResponse);
-
-            var persistedArticle = await _fixture.Context.Articles
-                .SingleAsync(a => a.Id == articleResponse.Id);
-
-            Assert.Equal(name, persistedArticle.Name);
-        }
-
-        [Fact]
-        public async Task Create_Article_With_IsRequired_Name_Attribute_Explicitly_Null_Fails()
-        {
-            // Arrange
-            var context = _fixture.GetService<AppDbContext>();
-            var author = _authorFaker.Generate();
-            context.AuthorDifferentDbContextName.Add(author);
-            await context.SaveChangesAsync();
-
-            var route = "/api/v1/articles";
-            var request = new HttpRequestMessage(new HttpMethod("POST"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    attributes = new Dictionary<string, object>
-                    {
-                        {"name", null}
-                    },
-                    relationships = new Dictionary<string, dynamic>
-                    {
-                        {  "author",  new
-                            {
-                                data = new
-                                {
-                                    type = "authors",
-                                    id = author.StringId
-                                }
-                            }
-                        }
-                    }
-                }
-            };
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.UnprocessableEntity, response.StatusCode);
-            var body = await response.Content.ReadAsStringAsync();
-            var errorDocument = JsonConvert.DeserializeObject<ErrorDocument>(body);
-            Assert.Single(errorDocument.Errors);
-            Assert.Equal("Input validation failed.", errorDocument.Errors[0].Title);
-            Assert.Equal("422", errorDocument.Errors[0].Status);
-            Assert.Equal("The Name field is required.", errorDocument.Errors[0].Detail);
-        }
-
-        [Fact]
-        public async Task Create_Article_With_IsRequired_Name_Attribute_Missing_Fails()
-        {
-            // Arrange
-            var context = _fixture.GetService<AppDbContext>();
-            var author = _authorFaker.Generate();
-            context.AuthorDifferentDbContextName.Add(author);
-            await context.SaveChangesAsync();
-
-            var route = "/api/v1/articles";
-            var request = new HttpRequestMessage(new HttpMethod("POST"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    relationships = new Dictionary<string, dynamic>
-                    {
-                        {  "author",  new
-                            {
-                                data = new
-                                {
-                                    type = "authors",
-                                    id = author.StringId
-                                }
-                            }
-                        }
-                    }
-                }
-            };
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.UnprocessableEntity, response.StatusCode);
-            var body = await response.Content.ReadAsStringAsync();
-            var errorDocument = JsonConvert.DeserializeObject<ErrorDocument>(body);
-            Assert.Single(errorDocument.Errors);
-            Assert.Equal("Input validation failed.", errorDocument.Errors[0].Title);
-            Assert.Equal("422", errorDocument.Errors[0].Status);
-            Assert.Equal("The Name field is required.", errorDocument.Errors[0].Detail);
-        } 
-
-        [Fact]
-        public async Task Update_Article_With_IsRequired_Name_Attribute_Succeeds()
-        {
-            // Arrange
-            var name = "Article Name";
-            var context = _fixture.GetService<AppDbContext>();
-            var article = _articleFaker.Generate();
-            context.Articles.Add(article);
-            await context.SaveChangesAsync();
-
-            var route = $"/api/v1/articles/{article.Id}";
-            var request = new HttpRequestMessage(new HttpMethod("PATCH"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    id = article.StringId,
-                    attributes = new Dictionary<string, object>
-                    {
-                        {"name", name}
-                    }
-                }
-            };
-
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            _fixture.ReloadDbContext();
-            var persistedArticle = await _fixture.Context.Articles
-                .SingleOrDefaultAsync(a => a.Id == article.Id);
-
-            var updatedName = persistedArticle.Name;
-            Assert.Equal(name, updatedName);
-        }
-
-        [Fact]
-        public async Task Update_Article_With_IsRequired_Name_Attribute_Missing_Succeeds()
-        {
-            // Arrange
-            var context = _fixture.GetService<AppDbContext>();
-            var tag = _tagFaker.Generate();
-            var article = _articleFaker.Generate();
-            context.Tags.Add(tag);
-            context.Articles.Add(article);
-            await context.SaveChangesAsync();
-
-            var route = $"/api/v1/articles/{article.Id}";
-            var request = new HttpRequestMessage(new HttpMethod("PATCH"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    id = article.StringId,
-                    relationships = new Dictionary<string, dynamic>
-                    {
-                        {  "tags",  new
-                            {
-                                data = new []
-                                {
-                                    new
-                                    {
-                                        type = "tags",
-                                        id = tag.StringId
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            };
-
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        }
-
-        [Fact]
-        public async Task Update_Article_With_IsRequired_Name_Attribute_Explicitly_Null_Fails()
-        {
-            // Arrange
-            var context = _fixture.GetService<AppDbContext>();
-            var article = _articleFaker.Generate();
-            context.Articles.Add(article);
-            await context.SaveChangesAsync();
-
-            var route = $"/api/v1/articles/{article.Id}";
-            var request = new HttpRequestMessage(new HttpMethod("PATCH"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    id = article.StringId,
-                    attributes = new Dictionary<string, object>
-                    {
-                        {"name", null}
-                    }
-                }
-            };
-
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.UnprocessableEntity, response.StatusCode);
-
-            var body = await response.Content.ReadAsStringAsync();
-            var errorDocument = JsonConvert.DeserializeObject<ErrorDocument>(body);
-            Assert.Single(errorDocument.Errors);
-            Assert.Equal("Input validation failed.", errorDocument.Errors[0].Title);
-            Assert.Equal("422", errorDocument.Errors[0].Status);
-            Assert.Equal("The Name field is required.", errorDocument.Errors[0].Detail);
-        }
-
-        [Fact]
-        public async Task Update_Article_With_IsRequired_Name_Attribute_Empty_Succeeds()
-        {
-            // Arrange
-            var context = _fixture.GetService<AppDbContext>();
-            var article = _articleFaker.Generate();
-            context.Articles.Add(article);
-            await context.SaveChangesAsync();
-
-            var route = $"/api/v1/articles/{article.Id}";
-            var request = new HttpRequestMessage(new HttpMethod("PATCH"), route);
-            var content = new
-            {
-                data = new
-                {
-                    type = "articles",
-                    id = article.StringId,
-                    attributes = new Dictionary<string, object>
-                    {
-                        {"name", ""}
-                    }
-                }
-            };
-
-            request.Content = new StringContent(JsonConvert.SerializeObject(content));
-            request.Content.Headers.ContentType = new MediaTypeHeaderValue(HeaderConstants.MediaType);
-
-            // Act
-            var response = await _fixture.Client.SendAsync(request);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            _fixture.ReloadDbContext();
-            var persistedArticle = await _fixture.Context.Articles
-                .SingleOrDefaultAsync(a => a.Id == article.Id);
-
-            var updatedName = persistedArticle.Name;
-            Assert.Equal("", updatedName);
-        }
     }
 }
